Make a form and redux boilerplates
Make a redux boileprate in index.js, connect store and redux dev tools.

Make an action creator that returns action that takes as a first
argument subjects's id, which is in config, and as the second config.
Action creator is called when one of pages is first rendered.

Make a toggleSubjectReducer that takes subject's id,
and assigns to it config object.

Make formik form boileprate.

Fix the bug that was in Header. Instead of component prop
in Route, I had render.